polyindex index with diskcache sharing
why? index has grew a lot with lots of data being duplicated.

the actual queries to pointed data does not happen often, hence things
can be moved to disk

furthermore, this allows to synchronise changes in the index

however the actual rtree are not shared, but only updated on
check_cache. alternative solution to this should be the postgis index,
that allows to share every aspect of the index